samples are now Double, added contiki code with new discovry formar

+++ cloud/conf/evolutions/default/1.sql
@@ -31,7 +31,7 @@ create table monitor (
 create table sample (
   resourceId                integer,
   timestamp                 integer,
-  value                     integer,
+  value                     double,
 );
 
 # --- !Downs
